refactor: move api wrappers; use "~" for deep (3 levels deep) components paths
