+++ __mocks__/server.ts
@@ -1,5 +1,5 @@
 import { setupServer } from 'msw/node'
-import { handlers } from '../sage/apis/ecr.mocks'
+import { handlers } from '../components/apis/ecr.mocks'
 
 // Setup requests interception using the given handlers.
 export const server = setupServer(...handlers)
