+++ admin-ui/views/status/charts/Charts.tsx
@@ -6,7 +6,7 @@ import 'chartjs-plugin-datalabels'
 import SummaryBar from './SummaryBar'
 import SummaryBox from './SummaryBox'
 
-import * as BK from '../../../apis/beekeeper'
+import * as BK from '../../../../components/apis/beekeeper'
 
 const barColors = {
   'reporting': '#3ac37e',
