+++ sage/ecr/apps/AppList.tsx
@@ -13,10 +13,10 @@ import PublicIcon from '@mui/icons-material/PublicRounded'
 import BeeIcon from 'url:../../../assets/bee.svg'
 
 import ErrorMsg from '../../ErrorMsg'
-import Table from '../../../components/table/Table'
-import TableSearch from '../../../components/table/TableSearch'
-import { useProgress } from '../../../components/progress/ProgressProvider'
-import * as ECR from '../../apis/ecr'
+import Table from '~/components/table/Table'
+import TableSearch from '~/components/table/TableSearch'
+import { useProgress } from '~/components/progress/ProgressProvider'
+import * as ECR from '~/components/apis/ecr'
 
 import {Top} from '../../common/Layout'
 import SpaciousLayout from './SpaciousAppList'
